move collection codecs to its own file
